gorilla/mux removed and replaced by http handling
