+++ server/server.go
@@ -3,47 +3,29 @@ package server
 import (
 	"net"
 	"net/http"
+	"net/http/httputil"
+	"net/url"
 	"os"
+	"regexp"
 	"strings"
 	"time"
 
 	"github.com/Sirupsen/logrus"
-	"github.com/gorilla/mux"
 	"github.com/spf13/viper"
+	"github.com/wemanity-belgium/hyperclair/database"
 	"github.com/wemanity-belgium/hyperclair/docker"
-	"github.com/wemanity-belgium/hyperclair/server/reverseProxy"
+	"github.com/wemanity-belgium/hyperclair/docker/httpclient"
 )
-
-type handler func(rw http.ResponseWriter, req *http.Request) error
-
-var router *mux.Router
 
 //Serve run a local server with the fileserver and the reverse proxy
 func Serve(sURL string) error {
 
 	go func() {
-		restrictedFileServer := func(path string) http.Handler {
-			if _, err := os.Stat(path); os.IsNotExist(err) {
-				os.Mkdir(path, 0777)
-			}
 
-			fc := func(w http.ResponseWriter, r *http.Request) {
-				http.FileServer(http.Dir(path)).ServeHTTP(w, r)
-			}
-			return http.HandlerFunc(fc)
-		}
+		http.Handle("/v2/", newSingleHostReverseProxy())
+		http.Handle("/local/", http.StripPrefix("/local", restrictedFileServer(docker.TmpLocal())))
 
-		router.PathPrefix("/v2/local").Handler(http.StripPrefix("/v2/local", restrictedFileServer(docker.TmpLocal()))).Methods("GET")
-		listener, err := net.Listen("tcp", sURL)
-		if err != nil {
-			logrus.Fatalf("cannot instanciate listener: %v", err)
-		}
-
-		if viper.GetInt("hyperclair.port") == 0 {
-			port := strings.Split(listener.Addr().String(), ":")[1]
-			logrus.Debugf("Update local server port from %q to %q", "0", port)
-			viper.Set("hyperclair.port", port)
-		}
+		listener := tcpListener(sURL)
 		logrus.Info("Starting Server on ", listener.Addr())
 
 		if err := http.Serve(listener, nil); err != nil {
@@ -55,9 +37,65 @@ func Serve(sURL string) error {
 	return nil
 }
 
-func init() {
+func tcpListener(sURL string) (listener net.Listener) {
+	listener, err := net.Listen("tcp", sURL)
+	if err != nil {
+		logrus.Fatalf("cannot instanciate listener: %v", err)
+	}
 
-	router = mux.NewRouter()
-	router.PathPrefix("/v2").Path("/{repository}/{name}/blobs/{digest}").HandlerFunc(reverseProxy.NewSingleHostReverseProxy().ServeHTTP)
-	http.Handle("/", router)
+	if viper.GetInt("hyperclair.port") == 0 {
+		port := strings.Split(listener.Addr().String(), ":")[1]
+		logrus.Debugf("Update local server port from %q to %q", "0", port)
+		viper.Set("hyperclair.port", port)
+	}
+
+	return
+}
+
+func restrictedFileServer(path string) http.Handler {
+	if _, err := os.Stat(path); os.IsNotExist(err) {
+		os.Mkdir(path, 0777)
+	}
+
+	fc := func(w http.ResponseWriter, r *http.Request) {
+		http.FileServer(http.Dir(path)).ServeHTTP(w, r)
+	}
+	return http.HandlerFunc(fc)
+}
+
+func newSingleHostReverseProxy() *httputil.ReverseProxy {
+	director := func(request *http.Request) {
+
+		var validID = regexp.MustCompile(`.*/blobs/(.*)$`)
+		u := request.URL.Path
+		logrus.Debugf("request url: %v", u)
+		if !validID.MatchString(u) {
+			logrus.Errorf("cannot parse url: %v", u)
+		}
+
+		host, _ := database.GetRegistryMapping(validID.FindStringSubmatch(u)[1])
+		out, _ := url.Parse(host)
+		request.URL.Scheme = out.Scheme
+		request.URL.Host = out.Host
+		client := httpclient.Get()
+		req, _ := http.NewRequest("HEAD", request.URL.String(), nil)
+		resp, err := client.Do(req)
+		if err != nil {
+			logrus.Errorf("response error: %v", err)
+			return
+		}
+
+		if resp.StatusCode == http.StatusUnauthorized {
+			logrus.Info("pull from clair is unauthorized")
+			docker.AuthenticateResponse(resp, request)
+		}
+
+		r, _ := http.NewRequest("GET", request.URL.String(), nil)
+		r.Header.Set("Authorization", request.Header.Get("Authorization"))
+		r.Header.Set("Accept-Encoding", request.Header.Get("Accept-Encoding"))
+		*request = *r
+	}
+	return &httputil.ReverseProxy{
+		Director: director,
+	}
 }
